requirements.txt
google-generativeai
langchain
langchain-google-genai
python-dotenv
streamlit